pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chiral"
version = "0.1.0"
description = "Compact WebSocket interface for robot policy evaluation"
requires-python = ">=3.10"
dependencies = [
  "websockets>=12.0",
  "numpy>=1.24",
  "msgpack>=1.0",
  "eclipse-zenoh>=1.7.2",
  "tqdm>=4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
docs = ["mkdocs-material>=9.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["chiral*"]
